Organize Excel enums into dedicated folder
