+++ OfficeIMO.Excel/Enums/NullPolicy.cs
@@ -1,0 +1,24 @@
+namespace OfficeIMO.Excel
+{
+    /// <summary>
+    /// Describes how null values are represented when flattening objects.
+    /// </summary>
+    public enum NullPolicy
+    {
+        /// <summary>
+        /// Replace nulls with an empty string.
+        /// </summary>
+        EmptyString,
+
+        /// <summary>
+        /// Keep null values as null references.
+        /// </summary>
+        NullLiteral,
+
+        /// <summary>
+        /// Use a default value supplied for the property.
+        /// </summary>
+        DefaultValue
+    }
+}
+
