+++ OfficeIMO.Excel/Enums/TableStyle.cs
@@ -1,0 +1,134 @@
+namespace OfficeIMO.Excel
+{
+    /// <summary>
+    /// Built-in table styles available in Excel.
+    /// </summary>
+    public enum TableStyle
+    {
+        // Light styles
+        /// <summary>Light style 1.</summary>
+        TableStyleLight1,
+        /// <summary>Light style 2.</summary>
+        TableStyleLight2,
+        /// <summary>Light style 3.</summary>
+        TableStyleLight3,
+        /// <summary>Light style 4.</summary>
+        TableStyleLight4,
+        /// <summary>Light style 5.</summary>
+        TableStyleLight5,
+        /// <summary>Light style 6.</summary>
+        TableStyleLight6,
+        /// <summary>Light style 7.</summary>
+        TableStyleLight7,
+        /// <summary>Light style 8.</summary>
+        TableStyleLight8,
+        /// <summary>Light style 9.</summary>
+        TableStyleLight9,
+        /// <summary>Light style 10.</summary>
+        TableStyleLight10,
+        /// <summary>Light style 11.</summary>
+        TableStyleLight11,
+        /// <summary>Light style 12.</summary>
+        TableStyleLight12,
+        /// <summary>Light style 13.</summary>
+        TableStyleLight13,
+        /// <summary>Light style 14.</summary>
+        TableStyleLight14,
+        /// <summary>Light style 15.</summary>
+        TableStyleLight15,
+        /// <summary>Light style 16.</summary>
+        TableStyleLight16,
+        /// <summary>Light style 17.</summary>
+        TableStyleLight17,
+        /// <summary>Light style 18.</summary>
+        TableStyleLight18,
+        /// <summary>Light style 19.</summary>
+        TableStyleLight19,
+        /// <summary>Light style 20.</summary>
+        TableStyleLight20,
+        /// <summary>Light style 21.</summary>
+        TableStyleLight21,
+
+        // Medium styles
+        /// <summary>Medium style 1.</summary>
+        TableStyleMedium1,
+        /// <summary>Medium style 2.</summary>
+        TableStyleMedium2,
+        /// <summary>Medium style 3.</summary>
+        TableStyleMedium3,
+        /// <summary>Medium style 4.</summary>
+        TableStyleMedium4,
+        /// <summary>Medium style 5.</summary>
+        TableStyleMedium5,
+        /// <summary>Medium style 6.</summary>
+        TableStyleMedium6,
+        /// <summary>Medium style 7.</summary>
+        TableStyleMedium7,
+        /// <summary>Medium style 8.</summary>
+        TableStyleMedium8,
+        /// <summary>Medium style 9.</summary>
+        TableStyleMedium9,
+        /// <summary>Medium style 10.</summary>
+        TableStyleMedium10,
+        /// <summary>Medium style 11.</summary>
+        TableStyleMedium11,
+        /// <summary>Medium style 12.</summary>
+        TableStyleMedium12,
+        /// <summary>Medium style 13.</summary>
+        TableStyleMedium13,
+        /// <summary>Medium style 14.</summary>
+        TableStyleMedium14,
+        /// <summary>Medium style 15.</summary>
+        TableStyleMedium15,
+        /// <summary>Medium style 16.</summary>
+        TableStyleMedium16,
+        /// <summary>Medium style 17.</summary>
+        TableStyleMedium17,
+        /// <summary>Medium style 18.</summary>
+        TableStyleMedium18,
+        /// <summary>Medium style 19.</summary>
+        TableStyleMedium19,
+        /// <summary>Medium style 20.</summary>
+        TableStyleMedium20,
+        /// <summary>Medium style 21.</summary>
+        TableStyleMedium21,
+        /// <summary>Medium style 22.</summary>
+        TableStyleMedium22,
+        /// <summary>Medium style 23.</summary>
+        TableStyleMedium23,
+        /// <summary>Medium style 24.</summary>
+        TableStyleMedium24,
+        /// <summary>Medium style 25.</summary>
+        TableStyleMedium25,
+        /// <summary>Medium style 26.</summary>
+        TableStyleMedium26,
+        /// <summary>Medium style 27.</summary>
+        TableStyleMedium27,
+        /// <summary>Medium style 28.</summary>
+        TableStyleMedium28,
+
+        // Dark styles
+        /// <summary>Dark style 1.</summary>
+        TableStyleDark1,
+        /// <summary>Dark style 2.</summary>
+        TableStyleDark2,
+        /// <summary>Dark style 3.</summary>
+        TableStyleDark3,
+        /// <summary>Dark style 4.</summary>
+        TableStyleDark4,
+        /// <summary>Dark style 5.</summary>
+        TableStyleDark5,
+        /// <summary>Dark style 6.</summary>
+        TableStyleDark6,
+        /// <summary>Dark style 7.</summary>
+        TableStyleDark7,
+        /// <summary>Dark style 8.</summary>
+        TableStyleDark8,
+        /// <summary>Dark style 9.</summary>
+        TableStyleDark9,
+        /// <summary>Dark style 10.</summary>
+        TableStyleDark10,
+        /// <summary>Dark style 11.</summary>
+        TableStyleDark11
+    }
+}
